chore(gulp): update gulp to 4 version, update all gulp dependencies, update all gulp tasks according to gulp 4, remove unused packages
