+++ package.json
@@ -125,7 +125,7 @@
     "@angular/language-service": "8.0.0",
     "@ng-bootstrap/ng-bootstrap": "^4.1.3",
     "@schematics/angular": "^8.0.0-rc.4",
-    "@types/gulp": "3.8.36",
+    "@types/gulp": "4.0.6",
     "@types/highlight.js": "9.12.2",
     "@types/jasmine": "2.8.3",
     "@types/jasminewd2": "2.0.3",
@@ -140,15 +140,13 @@
     "doc-prsr": "2.2.0",
     "firebase-tools": "^5.1.1",
     "fs-extra": "^8.1.0",
-    "gulp": "3.9.1",
-    "gulp-autoprefixer": "3.1.1",
-    "gulp-clean-css": "3.0.4",
-    "gulp-rename": "1.2.2",
-    "gulp-replace": "0.5.4",
+    "gulp": "4.0.2",
+    "gulp-rename": "1.4.0",
+    "gulp-replace": "1.0.0",
     "gulp-rollup": "2.13.0",
-    "gulp-sass": "3.1.0",
-    "gulp-sourcemaps": "2.6.0",
-    "gulp-typedoc": "2.2.1",
+    "gulp-sass": "4.0.2",
+    "gulp-sourcemaps": "2.6.5",
+    "gulp-typedoc": "2.2.3",
     "husky": "0.13.3",
     "jasmine-core": "2.8.0",
     "jasmine-spec-reporter": "4.2.1",
@@ -168,10 +166,9 @@
     "pump": "1.0.2",
     "rimraf": "2.6.1",
     "rollup-plugin-node-resolve": "3.0.0",
-    "run-sequence": "^1.2.2",
     "stylelint": "7.13.0",
     "through2": "^2.0.3",
-    "ts-node": "8.1.0",
+    "ts-node": "8.4.1",
     "tslint": "5.8.0",
     "typedoc": "^0.12.0",
     "typescript": "3.4.5",
