+++ scripts/gulp/gulpfile.ts
@@ -1,4 +1,3 @@
-import './tasks/build/build';
 import './tasks/build/resources';
 import './tasks/build/compile/scss';
 import './tasks/build/compile/typescript';
@@ -7,3 +6,4 @@ import './tasks/build/bundle/bundles-paths';
 import './tasks/bump-versions';
 import './tasks/change-prefix';
 import './tasks/docs/docs';
+import './tasks/build/build';
